requirements.txt
socket.io-emitter
mysql-replication